pyproject.toml
[project]
name = "bettercode"
version = "0.1.0"
description = "Code for BetterCodeBetterScience book"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter-book>=1.0.2",
    "numpy>=2.1.2",
    "scipy>=1.14.1",
    "jupyter>=1.1.1",
    "jupytext>=1.16.4",
    "pytest>=8.3.3",
    "ruff>=0.6.9",
    "pygithub>=2.4.0",
    "pandas>=2.2.3",
    "docutils==0.17.1",
    "pyppeteer>=2.0.0",
    "matplotlib>=3.9.2",
    "openai>=1.51.2",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "hypothesis>=6.115.3",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "transformers>=4.49.0",
    "torch>=2.6.0",
    "accelerate>=1.4.0",
    "icecream>=2.1.4",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "numba>=0.61,<0.63",
    "codespell>=2.4.1",
    "tomli>=2.2.1",
    "pre-commit>=4.2.0",
    "mdnewline>=0.1.3",
    "anthropic>=0.61.0",
    "nibabel>=5.3.2",
    "fastparquet>=2024.11.0",
    "templateflow>=25.1.1",
    "h5py>=3.15.1",
    "zarr>=3.1.3",
    "networkx>=3.4.2",
    "datalad>=1.2.3",
    "pickleshare>=0.7.5",
    "datalad-osf>=0.3.0",
    "pymongo[srv]>=4.15.4",
    "mysql-connector-python>=9.5.0",
    "biopython>=1.86",
    "neo4j>=6.0.3",
    "tqdm>=4.66.5",
    "fastembed>=0.7.3",
    "chromadb>=1.3.5",
    "monarch-py>=1.22.0",
    "biothings-client>=0.4.1",
    "gprofiler-official>=1.0.0",
    "ols-client>=0.2.1",
    "statsmodels>=0.14.5",
    "blue>=0.9.1",
    "nilearn>=0.12.1",
    "fmriprep-docker>=25.2.3",
    "mystmd>=1.7.0",
    "mne>=1.11.0",
    "mongomock>=4.3.0",
    "linkcheckmd>=1.4.0",
    "anndata>=0.12.7",
    "xarray>=2025.12.0",
    "dask>=2025.12.0",
    "pyarrow>=22.0.0",
    "scanpy>=1.11.5",
    "scrublet>=0.2.3",
    "igraph>=1.0.0",
    "leidenalg>=0.11.0",
    "fastcluster>=1.3.0",
    "scikit-misc>=0.5.2",
    "harmony-pytorch>=0.1.8",
    "pydeseq2>=0.5.3",
    "gseapy>=1.1.11",
    "ipython>=9.8.0",
    "harmonypy>=0.0.10",
    "rpy2>=3.6.4",
    "prefect>=3.0",
    "snakemake>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# add script entry points here
[project.scripts]
check-links = "bettercode.check_links:main"

[tool.codespell]
# Ref: https://github.com/codespell-project/codespell#using-a-config-file
skip = './book/transcripts,./book/_build/html,.js*,.git*,*.lock,*.bib,.venv*,*.ipynb,*.json'
check-hidden = true
ignore-regex = '\b(BLOSUM|Shapin)\b'
ignore-words = 'project-words.txt'

[tool.pytest.ini_options]
markers = [
    "unit: mark a test as a unit test.",
    "integration: mark a test as an integration test."
]
